Fallback to auto synth instead of silence if synth change failed (#15119)
Fixes #15096

Summary of the issue:
When an active profile has the synthesizer set to silence, and a user switches profiles to a synthesizer which fails to load, NVDA fails back to the silence synthesizer.

This is because the default behaviour on failure is to roll back to the previous synthesizer, which is appropriate except for silence.

As this is an attempt to change from silence to speech, instead NVDA should fallback to the next synthesizer, rather than the previous synthesizer.

Description of user facing changes
NVDA falls back to the silence synthesizer when:

a user is using the silence synthesizer,
and attempts to switch synthesizers e.g. via a profile change
and the synthesizer fails to load
Description of development approach
ensure the previous synthesizer we are switching back to is not silence.
Instead, the synth will switch to the next one, which may eventually reach silence if subsequent synthesizers fail to load.