Add-on store: Fix translations for add-ons installed from add-on store (#15152)
Closes #14973

Summary of the issue:
Add-ons installed from the add-on store used the untranslated add-on store JSON strings.
As of #15137, only add-ons that were installed from an external source had translated strings.

Description of user facing changes
Add-ons installed from the add-on store will have translated strings.

Description of development approach
Create a separate model for add-ons fetched from the add-on store, and add-ons cached after being installed from the add-on store.
Data fetched from the add-on store should be translated. Installed cached data for an add-on will be whatever language is used when the fetched add-on store data is cached. As such, we should defer to the translated manifests for installed add-ons.