Add option for ignoring blank lines for report line indentation (#15057)
closes #13394

Summary of the issue:
As elaborated in #13394, line indentation reporting can be too noisy in contexts where indentation changes for blank lines aren't meaningful and distract from the actually meaningful indentation changes, such as is the case when reading source code for Python and many other programming languages.

Description of user facing changes
A new checkbox is added to Document formatting, "Ignore blank lines for line indentation reporting". By default it is off, in which case line indentation reporting behaves as it does currently. If the user ticks the checkbox, then indentation changes are not reported for blank lines.

There is also a new unassigned global command for toggling the setting.

Description of development approach
new config property: ignoreBlankLinesForRLI = boolean(default=false)
short for ignoreBlankLinesForReportLineIndentation, which was too long and causing Flake8 issues
new settings panel checkbox for toggling ignoreBlankLinesForRLI
in speech.speech.getTextInfoSpeech, if ignoreBlankLinesForRLI is true and the line is blank, the block of code that inserts the indentation announcement and updates the indentation cache is skipped, otherwise it behaves as usual
user guide updated with brief info on new option
added unassigned global command for toggling ignoreBlankLinesForRLI
added system test for confirming the correct behavior