Add-on store: Fix up loading message (#15186)
Fixes #14975
Fixes #15184

Summary of the issue:
Sometimes the add-on store incorrectly reports "Loading add-ons" when it should report "No add-on selected".
This happens when no add-ons are found. When an add-on is found, the text is correctly replaced with the add-ons name.

When switching tabs, the add-on search filter should be reset.

Description of user facing changes
Fix bug where "No add-on selected" should be reported instead of "loading add-ons".

When switching tabs, the add-on search filter is now reset.

Description of development approach
Notify the details panel to update when the add-on list loading is complete. Previously, this would only refresh when an add-on was found.