Restore numpad gestures for move in flatten view of object hierarchy commands for all keyboard layouts (#15191)
Closes #15185
Follow-up of #15065.

Summary of the issue:
New gestures to move in flattened object hierarchy have been added in #15065: nvda+numpad3/9 for desktop layout and shift+NVDA+[ / ] for laptop layout.

Although mainly used with desktop layout, object nav-related gestures associated to numpad are usually bound to all keyboard layout (desktop and laptop). This allows for example users to have NVDA to be configured with laptop layout for a laptop computer without numpad on its keyboard, but to use the numpad gestures in case an external keyboard with numpad is plugged in.

Description of user facing changes
The gestures NVDA+numpad3/9 for the new flattened object navigation commands will now be available both in laptop and in desktop mode.

As done for other obj nav gestures, the user guide (and the change log) still advertises numpad gestures for desktop layout only.

Description of development approach
Changed the gesture definition.