braille.handler.setDisplayByName: fix logic error. (#15207)
Addresses issue introduced by pr #14524

Summary of the issue:
Reports from users that if a specifically configured braille display cannot be initialized on start-up, then the Braille category in the NVDA settings dialog is broken and does not show a valid braille driver (noBraille or otherwise).

Description of user facing changes
The Braille category in NVDA settings will continue to be usable if the configured Braille display cannot be initialized on start-up.

Description of development approach
In braille.handler.setDisplayByName if setting the display failed, the display is not set to NoBraille if the display was (not) already NoBraille. Previously this logic was incorrectly reversed.