Fixup #14924: again use weak references for functions (#15215)
Fixup of #14924

Summary of the issue:
In #14627, we introduced weak references for APCs called as part of a waitable timer. In #14924, this was made more robust by using a single internal APC func. However in the porting process, a part of the logic was reversed, therefore in the internal APC store, we still stored strong rather than weak references.

Description of user facing changes
None.

Description of development approach
Store references instead of functions in the apc store.