Always use utf-8 for addon store caches (#15241)
Closes #15231

Summary of the issue:
If NVDA language is set to Serbian, the cache could get corrupted while reading or writing do to the encoding not specified while opening the file.

Description of user facing changes
NVDA add-on store works with Serbian. Also, If the cached installed addon data is invalid, the exception will be logged similar to the add-on store cache, instead of preventing the add-on store from launching.

Description of development approach
open all files in _addonStore\dataManager.py in utf-8.
If the file is unable to be opened, or json.load fails, the exception will be logged.
Testing strategy:
Installed multiple add-ons, with Serbian translation and without, while NVDA's language is set to Serbian.