No longer classify (rich) edit classes as bad for UIA (#15314)
Fixes #14285

Summary of the issue:
There are reports of NVDA being sluggish in Rich Edit controls. In d5a3383bb2, rich edit classes were classified as bad for UIA for us to always use the win32 implementation. However, our support for UIA text editing is good enough for this restriction to be lifted, as this also solves the reported sluggishness.

Description of user facing changes
Using UIA instead of win32 API's under the hood.

Description of development approach
Moved the role, value and states overrides to a base class in window.edit
Rather than simply removing win32 edit classes when using UIA, insert the EditBase class as well. This ensures that role is reported as editable text rather than document, and that the multiline state is added. This ensures backwards compatibility as long as code authors don't match against MSAA specific properties.