Text alignment reporting and enhancements for MS Excel cells and MS Word text (#15205)
Closes #15206
Closes #15220

Summary of the issue:
Various issues for alignment reporting:

In MS Excel horizontal or vertical text alignment is not always correctly reported.
In MS Word (legacy), text alignment is not correctly reported for "justify" and for "distrubute".
Description of user facing changes
Reporting text alignment in Word and Excel is now handled better in various situations. "Default" is not reported anymore since it does not give any clear indication of the position of the text.

In MS Excel (no UIA) the alignment is reported as per Excel's options:
for horizontal text alignment in cell: "General", "Centered across columns", "Repeat" and "Distribute" is now reported instead of "Default" or nothing
for vertical alignment: "bottom" is reported instead of "default"
In MS Word (no UIA):
"justify" is now reported instead of "Default"
"distribute" is now reported instead of nothing
Description of development approach
Standardized alignments returned by getFormatField* functions around a string enum as done in #11897. Thus, all the interfaces are impacted.