Small BRLTTY driver improvements (#15335)
Partly related to issue #6483.

Summary of the issue:
The BRLTTY driver was available even when there was no BRLTTY instance with BrlAPI enabled running on the local computer (the only type of connection supported for now). Opening the actual connection has quite a long timeout, so this cannot be done as an availibility check. However, BrlAPI uses named pipes in a known location, so we can use that to check for availability. In the future we might also expose this as a port selection if multiple pipes are discovered and/or we add an interface to connect to a BrlAPI server on another host over TCP.

While I was working on this driver, I also marked it threadSafe and added a few more keybindings.

Description of user facing changes
The BRLTTY driver will not show up in the list of braille displays if no local BRLTTY with BrlAPI enabled is running.
More BRLTTY commands have been mapped to NVDA actions.