Fixes to address hangs caused by WASAPI idle checks (#15349)
Fixes #15150.

Summary of the issue:
When WASAPI is enabled, NVDA sometimes hangs and cannot be recovered, especially when switching audio devices. The freezes all occur in WasapiWavePlayer._idleCheck when calling wasPlay_idle.

Description of user facing changes
Fixed the hangs.

Description of development approach
If getting the playback position fails with an error, playback has probably been interrupted; e.g. because the device disconnected. Treat this as if playback has finished so we don't wait forever and so that we fire any pending callbacks.
Never treat a stream as idle while we're feeding audio to it. Aside from this not making any sense, we might otherwise try to sync and stop it in the main thread, which could cause hangs or crashes.