Fix WASAPI option enabling/disabling when restoring advanced options (#15355)
Summary of the issue:
The following advanced options are greyed out or not depending if they are applicable: "Volume of NVDA sounds" and "Volume of NVDA sounds follows voice volume". If they are applicable or not depends on other options: "Use WASAPI for audio output" and "Volume of NVDA sounds" respectively. Unfortunately, when pressing the "Restore defaults" button of the advanced panel, the options are reset to there default values, but they are not greyed out / ungreyed accordingly.

Description of user facing changes
When pressing the restore default button in advanced panel, WASAPI advanced options will now be greyed out or not according to if they are applicable or not.

Description of development approach
Just call the update function self._onWASAPIChange() when restoring the values.