Reintroduce Improve shortcut key reporting (#14900) (#15373)
Fixes #10807
Reintroduce the changes of PR #14900 that were reverted in #15237.

Summary of the issue:
See #14900 for the initial issue.

#14900 had to be reverted since it caused any speech sequence containing a character mode command to spell the whole speech sequence with Code Factory driver. Although the issue is in the driver and not NVDA, it has been decided to wait for an API-breaking release to issue this fix so that Code Factory synth users be not impacted by this bug.

Description of user facing changes
Same as #14900, that is:

Shortcut key reporting will be improved in some cases. This applies to on-demand reporting (shift+numpad2) as well as when "Report Object Shortcut Keys" option is checked in Object presentation settings.
The cases where it should be improved are:

punctuation key, e.g. "." (dot)
letter keys when the name of the letter is not spoken the same way as the equivalent one-letter word, e.g. "à" or "y" in French.
Braille reporting should remain unchanged.
Description of development approach
Revert #15237 to reintroduce the changes in #14900 and remove change log modifications.