Remove `SecureDesktopNVDAObject`, replace with extension point (#14488)
Fixes #14137

Summary of the issue:
SecureDesktopNVDAObject is used to notify the user and API consumers that NVDA has entered a secure desktop.
This creates a valid NVDAObject, backed by the secure desktop.
The running instance of NVDA enters sleep mode when SecureDesktopNVDAObject is focused, signifying a switch to a secure desktop. Then a new instance of NVDA in secure mode starts on the secure desktop.
The SecureDesktopNVDAObject is unused, apart from being an API endpoint to notify consumers, as NVDA sleeps after it is created.

Handling SecureDesktopNVDAObject requires special cases to be constructed.

NVDA generally uses extension points for this API use case.

Description of user facing changes
None

Description of development approach
SecureDesktopNVDAObject is replaced with an extension point that only provides a boolean state on the existence of the secure desktop.
A fake NVDAObject is used to enable sleep mode while the secure desktop is active.