Documentation fix-up for new winAPI.secureDesktop.post_secureDesktopStateChange extensionPoint (#15392)
Fix-up of #14488

Summary of the issue:
In #14488, no documentation for the new extensionPoint was added to the master list of extension points in the developer guide.
(I wish there was a way to remind of this when contributing extensionPoint creations)

Additionally, I noticed an oddity in the documentation of the extensionPoint itself, which may or may not be intended.

Description of user facing changes
Improved documentation for developers.

Description of development approach
Added a new section to the Developer Guide chapter on Extension Points, listing the new Action, winAPI.secureDesktop.post_secureDesktopStateChange.
Fixed a terminology error in the head section of the chapter (I had previously called extensionPoints a module, when it should have been a package).
Noticed that in the usage example given in the winAPI.secureDesktop module docstring, there was a call to notify(). AFAIK, the general consumer of this extensionPoint, who registers something to it, shouldn't also be calling notify on it. I removed that example line, replacing it with a comment to put the subsequent unregister() call in context. @seanbudd please check my reasoning.