Add checks for region updates (#15393)
Fixes #15391
Fixup of #15163

Summary of the issue:
Since #15163, it is much more likely that braille tries to update a region for an object that no longer exists. This is no problem at all, however in that case, the region should simply be ignored for updating and a debug warning should be logged instead.

Description of user facing changes
No errors in de log while browsing

Description of development approach
a try/except around region.update, with an extra check to avoid updating regions for tree interceptors that have died.

Testing strategy:
Browsed with Firefox for some minutes, wasn't able to reproduce the issue any longer.