Align horizontally checkbox in add-on store GUI (#15428)
Fixes #15018

Summary of the issue:
In the add-on store, the "include incompatible add-ons" checkbox is not aligned with the combobox at its left.

Description of user facing changes
The checkbox will be aligned horizontally.

Description of development approach
In the guiHelper, added the wx.EXPAND flag so that it is centered vertically in the sizer.

I do not know if other items may benefit from this flag; we could add them in the if clause if / when needed in the future. such Horizontal sizers are less common in NVDA's GUI however.