Force incompatibility for NVDAExtensionGlobalPlugin versions 12.0.8 or below (#15443)
Follow-up of #15402

See #15402 (comment)

Summary of the issue:
As with Tony's Enhancements add-on, "NVDA global commands extension" add-on (name = NVDAExtensionGlobalPlugin) version 12.0.8 or below causes speech muted with WASAPI, so with NVDA 2023.3beta1.

This add-on is not in NVDA store but is popular in French community as well as in international community; it is translated in 10 languages.

A fix release (13.0) has already been created by the author @paulber19 on August 20. But we cannot ensure that people will have updated before updating NVDA.

Description of user facing changes
No problem when running NVDA with NVDAExtensionGlobalPlugin 12.0.8: NVDA is speaking.

Description of development approach
Added this add-on and its version in the list to force incompatibility.