Disable "Ignore blank lines" checkbox when report indent is off (#15494)
Follow-up of #15057

Summary of the issue:
In the GUI, the recent trend is to disable (grey out) the options that are not applicable due to the values of other options.
New option "Ignore blank lines for line indentation reporting" has no sense when report indentation is OFF. Thus it should be greyed out when report indent is off.

This is especially needed to reduce the number of tab presses when navigating in the Doc formatting settings dialog which contains a lot of items.

Description of user facing changes
Option "Ignore blank lines for lineis greyed out when report indent is off and enabled when other indent reporting types are configured.

While at it, in this PR, I have also changed the context help target for "Ignore blank lines" option. Before it was jumping by default to "Document formatting settings" paragraph; now it jumps to "Line indentation reporting" paragraph, which contains more precise information on this topic.