Depend on supported command for capital pitch change (#15433)
Closes #15262
Blocked by #15271

Summary of the issue:
When implementing a synthesizer driver, there are two ways to change pitch:

Change the pitch setting
Send a pitch command that changes the pitch for the duration of the sequence or until there is a new pitch command.
Capital pitch change uses pitch commands to change pitch for capital letters. When doing this, it calls isSupported("pitch") on the driver, which returns True when the pitch setting is supported. However in order to support capital pitch changes, the driver should implement support for the PitchCommand instead.

Description of user facing changes
None

Description of development approach
Changed all situations where isSupported("pitch") is called to generate a pitch command to check whether the PitchCommand is in supportedCommands.