Add spelling for double presses of some NVDA commands (#15504)
Closes #15449

Summary of the issue:
Various NVDA commands which report information spells it on second press. This was not the case for the two following commands:

Report current selection
Report clipboard text
Moreover, the command to report focus (NVDA+Tab) support spelling the information on second press, but not the triple press version spelling it using character description.

Description of user facing changes
When pressing two times NVDA+C or NVDA+shift+upArrow, the reported information will be spelt. A third press spells the information using character description. As for other similar scripts, if the selection or the clipboard text contains too many characters, the information is not spelt and the number of characters is reported instead.

Similarly, the command to report focus (NVDA+Tab) has also been extended to support spelling the information using character description on third press.