Update SCons to version 4.5.2 (#15529)
Summary of the issue:
As part of the investment to update all NVDA dependencies, SCons needs an update to the most recent version.

Description of user facing changes
Updates Scons to 4.5.2

Description of development approach
Updates requirements. As CPPDEFINES is now saved as a deque internally, I needed to convert the deque to a list before expanding it for one of the libraries.