Add-on store: Handle bulk install (#15350)
Summary of the issue:
To be able to stress test the add-on installation system, a method of bulk installing add-ons is useful.
There is also a general request from users to be able to perform bulk actions like installing, updating or removing add-ons.

Description of user facing changes
Users are now able to bulk install add-ons.
This can be done by selecting multiple add-ons in the available add-ons tab, then activating the context menu on the selection and choosing the install action.

Description of development approach
Create a separate context menu for bulk actions. This creates an API for future bulk actions.