Sapi4: Only send prosody commands when there is prosody in the sequence (#15583)
Fixes #15580
Follow up of #15502

Summary of the issue:
In #15502, I fixed some regressions regarding capital pitch changes introduced in #15271. This meant that we'd add prosody commands to every speech sequence to support a specific version of IBM TTS that seemed to need that. However, it introduced a regression in that it was no longer possible to change speech parameters using the gui.

Description of user facing changes
It is again possible to change SAPI4 parameters using the GUI.

Description of development approach
Only add blank prosody commands when there is any prosody in the sequence. IF not, don't do anything.